+++ experimental/packages/sdk-logs/src/LoggerProvider.ts
@@ -17,12 +17,14 @@ import { diag } from '@opentelemetry/api';
 import type * as logsAPI from '@opentelemetry/api-logs';
 import { NOOP_LOGGER } from '@opentelemetry/api-logs';
 import { defaultResource } from '@opentelemetry/resources';
-import { BindOnceFuture, merge } from '@opentelemetry/core';
+import { BindOnceFuture } from '@opentelemetry/core';
 
 import type { LoggerProviderConfig } from './types';
 import { Logger } from './Logger';
-import { loadDefaultConfig, reconfigureLimits } from './config';
-import { LoggerProviderSharedState } from './internal/LoggerProviderSharedState';
+import {
+  DEFAULT_LOGGER_CONFIGURATOR,
+  LoggerProviderSharedState,
+} from './internal/LoggerProviderSharedState';
 
 export const DEFAULT_LOGGER_NAME = 'unknown';
 
@@ -31,14 +33,24 @@ export class LoggerProvider implements logsAPI.LoggerProvider {
   private readonly _sharedState: LoggerProviderSharedState;
 
   constructor(config: LoggerProviderConfig = {}) {
-    const mergedConfig = merge({}, loadDefaultConfig(), config);
-    const resource = config.resource ?? defaultResource();
+    const mergedConfig = {
+      resource: config.resource ?? defaultResource(),
+      forceFlushTimeoutMillis: config.forceFlushTimeoutMillis ?? 30000,
+      logRecordLimits: {
+        attributeCountLimit: config.logRecordLimits?.attributeCountLimit ?? 128,
+        attributeValueLengthLimit:
+          config.logRecordLimits?.attributeValueLengthLimit ?? Infinity,
+      },
+      loggerConfigurator:
+        config.loggerConfigurator ?? DEFAULT_LOGGER_CONFIGURATOR,
+      processors: config.processors ?? [],
+    };
     this._sharedState = new LoggerProviderSharedState(
-      resource,
+      mergedConfig.resource,
       mergedConfig.forceFlushTimeoutMillis,
-      reconfigureLimits(mergedConfig.logRecordLimits),
-      config?.processors ?? [],
-      config?.loggerConfigurator
+      mergedConfig.logRecordLimits,
+      mergedConfig.processors,
+      mergedConfig.loggerConfigurator
     );
     this._shutdownOnce = new BindOnceFuture(this._shutdown, this);
   }